+++ libraries/transformer/src/main/java/androidx/media3/transformer/FrameworkMuxer.java
@@ -83,6 +83,7 @@ import java.nio.ByteBuffer;
   private final long videoDurationUs;
   private final MediaCodec.BufferInfo bufferInfo;
   private final SparseLongArray trackIndexToLastPresentationTimeUs;
+  private final SparseLongArray trackIndexToPresentationTimeOffsetUs;
 
   private int videoTrackIndex;
 
@@ -95,6 +96,7 @@ import java.nio.ByteBuffer;
     this.videoDurationUs = Util.msToUs(videoDurationMs);
     bufferInfo = new MediaCodec.BufferInfo();
     trackIndexToLastPresentationTimeUs = new SparseLongArray();
+    trackIndexToPresentationTimeOffsetUs = new SparseLongArray();
     videoTrackIndex = C.INDEX_UNSET;
   }
 
@@ -145,6 +147,9 @@ import java.nio.ByteBuffer;
 
     if (!isStarted) {
       isStarted = true;
+      if (Util.SDK_INT < 30 && presentationTimeUs < 0) {
+        trackIndexToPresentationTimeOffsetUs.put(trackIndex, -presentationTimeUs);
+      }
       try {
         mediaMuxer.start();
       } catch (RuntimeException e) {
@@ -154,6 +159,9 @@ import java.nio.ByteBuffer;
 
     int offset = data.position();
     int size = data.limit() - offset;
+
+    long presentationTimeOffsetUs = trackIndexToPresentationTimeOffsetUs.get(trackIndex);
+    presentationTimeUs += presentationTimeOffsetUs;
 
     bufferInfo.set(offset, size, presentationTimeUs, TransformerUtil.getMediaCodecFlags(flags));
     long lastSamplePresentationTimeUs = trackIndexToLastPresentationTimeUs.get(trackIndex);
@@ -166,6 +174,15 @@ import java.nio.ByteBuffer;
             + lastSamplePresentationTimeUs
             + ") unsupported on this API version");
     trackIndexToLastPresentationTimeUs.put(trackIndex, presentationTimeUs);
+
+    checkState(
+        presentationTimeOffsetUs == 0 || presentationTimeUs >= lastSamplePresentationTimeUs,
+        "Samples not in presentation order ("
+            + presentationTimeUs
+            + " < "
+            + lastSamplePresentationTimeUs
+            + ") unsupported when using negative PTS workaround");
+
     try {
       mediaMuxer.writeSampleData(trackIndex, data, bufferInfo);
     } catch (RuntimeException e) {